pyproject.toml
[project]
name = "src-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "minio>=7.2.16",
    "pillow>=11.3.0",
    "pillow-jxl-plugin>=1.3.4",
    "python-dotenv>=1.1.1",
]
